requirements.txt
psycopg2-binary
coverage
pyhdb
black
flake8
isort
torch
gensim==3.8.3
numpy==1.19.4
opencv-python==4.4.0.46
pandas==1.1.4
tensorboard==1.15.0
tensorflow==1.15.0
lark
cloudpickle==1.6.0
dppy==0.3.2
joblib==1.2.0
line_profiler==4.0.2
matplotlib==3.5.3
mpi4py==3.1.4
mujoco_py==2.1.2.14
scikit_learn==1.0.2
scipy==1.7.3
seaborn==0.13.0
sphinx_rtd_theme==1.3.0
stable_baselines3==2.1.0